+++ init.lua
@@ -5,11 +5,11 @@ local root = require "PLoop.Prototype"
 require "PLoop.System.Collections"
 require "PLoop.System.Text"
 require "PLoop.System.IO.TextAccessor"
+require "PLoop.System.Threading"
 require "PLoop.System.Serialization"
 require "PLoop.System.Date"
 require "PLoop.System.Logger"
 require "PLoop.System.Recycle"
-require "PLoop.System.Threading"
 require "PLoop.System.Observer"
 
 -- System.Collections